chore(generator): rename package 'arch' to 'archunit'
