+++ src/main/resources/generator/server/javatool/archunit/AnnotationArchTest.java.mustache
