+++ src/main/resources/generator/server/javatool/archunit/HexagonalArchTest.java.mustache
