+++ src/main/resources/generator/server/javatool/archunit/archunit.properties.mustache
